+++ src/runtime_src/hip/api/CMakeLists.txt
@@ -6,11 +6,11 @@ add_library(hip_api_library_objects OBJECT
   hip_module.cpp
   hip_stream.cpp
   hip_memory.cpp
+  hip_event.cpp
 )
 
 target_include_directories(hip_api_library_objects
   PRIVATE
   ${XRT_SOURCE_DIR}/runtime_src
-  ${XRT_SOURCE_DIR}/runtime_src/core/include
   ${HIP_INCLUDE_DIRS}
   )
