--- src/runtime_src/hip/core/buffer_copy.cpp
@@ -1,18 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (C) 2024 Advanced Micro Device, Inc. All rights reserved.
-
-#include "buffer_copy.h"
-
-namespace xrt::core::hip {
-
-bool
-buffer_copy::
-submit()
-{}
-
-bool
-buffer_copy::
-wait()
-{}
-
-}
