--- src/runtime_src/hip/core/buffer_copy.h
@@ -1,24 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (C) 2024 Advanced Micro Device, Inc. All rights reserved.
-#ifndef xrthip_buffer_copy_h
-#define xrthip_buffer_copy_h
-
-#include "command.h"
-
-namespace xrt::core::hip {
-
-class copy_buffer : public command
-{
-private:
-  std::shared_ptr<buffer> buff;
-  direction cdirection;
-
-public:
-  bool submit();
-  bool wait();
-
-};
-
-} // xrt::core::hip
-
-#endif
