--- src/runtime_src/hip/core/command.cpp
@@ -1,18 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (C) 2024 Advanced Micro Device, Inc. All rights reserved.
-
-#include "command.h"
-
-namespace xrt::core::hip {
-
-bool
-command::
-submit()
-{}
-
-bool
-command::
-wait()
-{}
-
-}
