--- src/runtime_src/hip/core/kernel_start.cpp
@@ -1,22 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (C) 2024 Advanced Micro Device, Inc. All rights reserved.
-
-#include "kernel_start.h"
-
-namespace xrt::core::hip {
-
-kernel_start::
-kernel_start(xrt::kernel &, void* args)
-{}
-
-bool
-kernel_start::
-submit()
-{}
-
-bool
-kerel_start::
-wait()
-{}
-
-}
